Modify noise handler to an ask-tell architecture using a moving avergage predictor to provide the expected noise for the current step.
Mutations for the noise handler calculation of the previous step are done in the current step and all simulations are now handled in a single queue.

Solutions are stored alongside function values in the result queue to ensure that the do not get jumbled.

If the noise handler reaches max evalutions sigma is no longer increased.